ChangeLog (20260518-1917): Backend: 1) disable the default spring security in memory user(stop generating default login user/password) 2) Add in new endpoint for getSlotBookedTimesBySlotIdV1, remove countBySlotIdV1 at BookingController, getSlotBookedTimesBySlotId will be used to feed info to frontend, for guest usage, avoid sensitive infomation, only info on bookedstarttime and bookedendtime is provided 3)  Add in new endpoint eventDeleteValidationV1 at EventController , Add in new endpoint slotDeleteValidationV1 at SlotController, this is to implement guard to prevent organizer delete event or slot when there is active booking on going. User are require to cancel the related active booking before can remove the event or slot 4) Modify existing deleteEventById at EventService and existing deleteSlotByEventAndSlot at SlotService to validate activebooking before delete can be done. 5) Update initGuestBookingViewAccess at GuestBookingAccessService, now valid will be null if captchaRequired is true, this is to prevent user to access valid field (with developmentootl) without doing the captcha. 6) Adding in check bookingstatus method in BookingStatus class. 7) Update repo that work abnormally after setting spring.jpa.open-in-view=false 8) Update GuestPrincipal to include bookingtoken, which will be used to build guest authentication in JwtAuthenticationProvider, this increase security when guest is trying to fetch booking info, the bookingtoken info is encoded in JWT Token and later use by service to fetch booking info. 9) Update the path that can be permitall at SecurityConfig and JwtAuthFilter 10) Added in EmailService method: sendBookingCancelledDetails and implement it related delete booking service method 11) Move some of the business logic away from dtomapper 12) BusinessDaysHours now support frontend to send in time field with 24:00, indicate as end of the day 2400
Frontend:
1) Created LoggerService, implement logging function from LoggerService at classes and components
2)  Create ErrorCode, ErrorMapperService and updated error interceptor flow. Alert message will be send via error interceptor base on the ErrorMapperService and GlobalErrorService now.
3) Update Full Calendar View,
a) implemented visibleRange when creating calendar event.
b) Refactor code by creating refreshCalendarData() which will be use to refreshCalendarData when necessary especially when changing view between month, week and day.
c) Businees event has refactor to not use recurrent event, instead real date event is use to avoid timezone changing bug, with the implementation of visibleRange, real date event can be render without causing high load crashing issue.
d) Calendar intial view option is refactor to only apply once every component is generated, and will not re trigger when refreshcalendardata is called.

+++ 02-backend-booking-system/src/main/java/com/littlek4za/booking_system/services/SlotServiceImpl.java
@@ -8,8 +8,10 @@ import java.util.stream.Collectors;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
+import com.littlek4za.booking_system.dtos.DeleteValidationResponseDto;
 import com.littlek4za.booking_system.dtos.SlotRequestDto;
 import com.littlek4za.booking_system.dtos.SlotResponseDto;
+import com.littlek4za.booking_system.entities.Booking;
 import com.littlek4za.booking_system.entities.Event;
 import com.littlek4za.booking_system.entities.Invitation;
 import com.littlek4za.booking_system.entities.Slot;
@@ -39,11 +41,13 @@ public class SlotServiceImpl implements SlotService {
         private final SlotValidator slotValidator;
         private final BookingRepository bookingRepository;
         private final InvitationRepository invitationRepository;
+        private final DeleteValidationService deleteValidationService;
 
         public SlotServiceImpl(SecurityUtil sercurityUtil, SlotRepository slotRepository,
                         EventRepository eventRepository,
                         UserRepository userRepository, DtoMapper dtoMapper, SlotValidator slotValidator,
-                        BookingRepository bookingRepository, InvitationRepository invitationRepository) {
+                        BookingRepository bookingRepository, InvitationRepository invitationRepository,
+                        DeleteValidationService deleteValidationService) {
                 this.securityUtil = sercurityUtil;
                 this.slotRepository = slotRepository;
                 this.eventRepository = eventRepository;
@@ -52,21 +56,24 @@ public class SlotServiceImpl implements SlotService {
                 this.slotValidator = slotValidator;
                 this.bookingRepository = bookingRepository;
                 this.invitationRepository = invitationRepository;
+                this.deleteValidationService = deleteValidationService;
         }
 
         @Override
         public List<SlotResponseDto> getSlotsByEventId(Long eventId) {
 
                 User user = userRepository.findById(this.securityUtil.requireUserId())
-                                .orElseThrow(() -> new AppException("User not found", HttpStatus.NOT_FOUND, ErrorCode.USER_NOT_FOUND));
+                                .orElseThrow(() -> new AppException("User not found", HttpStatus.NOT_FOUND,
+                                                ErrorCode.USER_NOT_FOUND));
                 Event event = eventRepository.findByIdAndUser(eventId, user)
-                                .orElseThrow(() -> new AppException("Event not found with eventId and user", HttpStatus.NOT_FOUND, ErrorCode.EVENT_NOT_FOUND));
+                                .orElseThrow(() -> new AppException("Event not found with eventId and user",
+                                                HttpStatus.NOT_FOUND, ErrorCode.EVENT_NOT_FOUND));
                 Map<Long, Long> bookingsCountMap = this.bookingRepository.countBookingsByEventGrouped(eventId)
-                                                        .stream()
-                                                        .collect(Collectors.toMap(
-                                                                row -> (Long) row[0],
-                                                                row -> (Long) row[1]));
-                Set<Slot> slotSet = slotRepository.findByEvent(event);
+                                .stream()
+                                .collect(Collectors.toMap(
+                                                row -> (Long) row[0],
+                                                row -> (Long) row[1]));
+                Set<Slot> slotSet = slotRepository.findByEventWithEvent(event);
                 List<SlotResponseDto> slotResponseDtoList = slotSet.stream()
                                 .map(slot -> {
                                         Long bookingsCount = bookingsCountMap.getOrDefault(slot.getId(), 0L);
@@ -77,14 +84,16 @@ public class SlotServiceImpl implements SlotService {
                 return slotResponseDtoList;
         }
 
-        @Transactional
         @Override
+        @Transactional
         public SlotResponseDto createSlotByEvent(Long eventId, SlotRequestDto slotRequestDto) {
 
                 User user = userRepository.findById(this.securityUtil.requireUserId())
-                                .orElseThrow(() -> new AppException("User not found", HttpStatus.NOT_FOUND, ErrorCode.USER_NOT_FOUND));
+                                .orElseThrow(() -> new AppException("User not found", HttpStatus.NOT_FOUND,
+                                                ErrorCode.USER_NOT_FOUND));
                 Event event = eventRepository.findByIdAndUser(eventId, user)
-                                .orElseThrow(() -> new AppException("Event not found with eventId and user", HttpStatus.NOT_FOUND, ErrorCode.EVENT_NOT_FOUND));
+                                .orElseThrow(() -> new AppException("Event not found with eventId and user",
+                                                HttpStatus.NOT_FOUND, ErrorCode.EVENT_NOT_FOUND));
 
                 slotValidator.validate(event.getEventType(), slotRequestDto);
 
@@ -105,15 +114,34 @@ public class SlotServiceImpl implements SlotService {
         @Transactional
         public Long deleteSlotByEventAndSlot(Long eventId, Long slotId) {
                 User user = userRepository.findById(securityUtil.requireUserId())
-                                .orElseThrow(() -> new AppException("User not found", HttpStatus.NOT_FOUND, ErrorCode.USER_NOT_FOUND));
+                                .orElseThrow(() -> new AppException("User not found", HttpStatus.NOT_FOUND,
+                                                ErrorCode.USER_NOT_FOUND));
                 Event event = eventRepository.findByIdAndUser(eventId, user)
-                                .orElseThrow(() -> new AppException("Event not found with eventId and user", HttpStatus.NOT_FOUND, ErrorCode.EVENT_NOT_FOUND));
-                // handle invitation deletion                                
+                                .orElseThrow(() -> new AppException("Event not found with eventId and user",
+                                                HttpStatus.NOT_FOUND, ErrorCode.EVENT_NOT_FOUND));
+                // handle invitation deletion
                 Slot slot = slotRepository.findByIdAndEventIdWithInvitationSet(slotId, event.getId())
-                                .orElseThrow(() -> new AppException("Slot not found with slotId and eventId", HttpStatus.NOT_FOUND, ErrorCode.SLOT_NOT_FOUND));
-                for(Invitation invitation : slot.getInvitationSet()){
-                        if(invitation.getSlotIncludeMode() == SlotIncludeMode.SELECTED){
-                                if(invitation.getSlotSet().size() == 1) {
+                                .orElseThrow(() -> new AppException("Slot not found with slotId and eventId",
+                                                HttpStatus.NOT_FOUND, ErrorCode.SLOT_NOT_FOUND));
+
+                List<Booking> bookingList = bookingRepository.findBySlot_IdAndSlot_Event_IdAndIsDeletedFalse(slotId, eventId);
+
+                boolean canDelete = true;
+
+                if (!bookingList.isEmpty()) {
+                        canDelete = deleteValidationService.buildDeleteValidationResponseDto(bookingList).canDelete();
+                }
+
+                if (canDelete == false) {
+                        throw new AppException(
+                                        "Slot cannot be deleted due to active bookings",
+                                        HttpStatus.BAD_REQUEST,
+                                        ErrorCode.SLOT_HAS_ACTIVE_BOOKINGS);
+                }
+
+                for (Invitation invitation : slot.getInvitationSet()) {
+                        if (invitation.getSlotIncludeMode() == SlotIncludeMode.SELECTED) {
+                                if (invitation.getSlotSet().size() == 1) {
                                         invitationRepository.delete(invitation);
                                 } else {
                                         invitation.getSlotSet().remove(slot);
@@ -125,7 +153,8 @@ public class SlotServiceImpl implements SlotService {
                 int deleted = this.slotRepository.deleteByIdAndEventId(slotId, event.getId());
 
                 if (deleted == 0) {
-                        throw new AppException("Slot not found with slotId and eventId", HttpStatus.NOT_FOUND, ErrorCode.SLOT_NOT_FOUND);
+                        throw new AppException("Slot not found with slotId and eventId", HttpStatus.NOT_FOUND,
+                                        ErrorCode.SLOT_NOT_FOUND);
                 }
 
                 return slotId;
@@ -134,26 +163,33 @@ public class SlotServiceImpl implements SlotService {
         @Override
         public SlotResponseDto getSlotByIdAndEventId(Long eventId, Long slotId) {
                 User user = userRepository.findById(securityUtil.requireUserId())
-                                .orElseThrow(() -> new AppException("User not found", HttpStatus.NOT_FOUND, ErrorCode.USER_NOT_FOUND));
+                                .orElseThrow(() -> new AppException("User not found", HttpStatus.NOT_FOUND,
+                                                ErrorCode.USER_NOT_FOUND));
                 Event event = eventRepository.findByIdAndUser(eventId, user)
-                                .orElseThrow(() -> new AppException("Event not found with eventId and user", HttpStatus.NOT_FOUND, ErrorCode.EVENT_NOT_FOUND));
+                                .orElseThrow(() -> new AppException("Event not found with eventId and user",
+                                                HttpStatus.NOT_FOUND, ErrorCode.EVENT_NOT_FOUND));
                 Slot slot = slotRepository.findByIdAndEventId(slotId, event.getId())
-                                .orElseThrow(() -> new AppException("Slot not found with slotId and eventId", HttpStatus.NOT_FOUND, ErrorCode.SLOT_NOT_FOUND));
-                
+                                .orElseThrow(() -> new AppException("Slot not found with slotId and eventId",
+                                                HttpStatus.NOT_FOUND, ErrorCode.SLOT_NOT_FOUND));
+
                 Long bookingsCount = bookingRepository.countBySlotId(slotId);
 
                 return dtoMapper.toSlotResponseDto(slot, bookingsCount);
         }
 
         @Override
+        @Transactional
         public SlotResponseDto putSlotByIdAndEventId(Long slotId, Long eventId, SlotRequestDto slotRequestDto) {
                 User user = userRepository.findById(securityUtil.requireUserId())
-                                .orElseThrow(() -> new AppException("User not found", HttpStatus.NOT_FOUND, ErrorCode.USER_NOT_FOUND));
+                                .orElseThrow(() -> new AppException("User not found", HttpStatus.NOT_FOUND,
+                                                ErrorCode.USER_NOT_FOUND));
                 Event event = eventRepository.findByIdAndUser(eventId, user)
-                                .orElseThrow(() -> new AppException("Event not found with eventId and user", HttpStatus.NOT_FOUND, ErrorCode.EVENT_NOT_FOUND));
+                                .orElseThrow(() -> new AppException("Event not found with eventId and user",
+                                                HttpStatus.NOT_FOUND, ErrorCode.EVENT_NOT_FOUND));
 
                 Slot existingSlot = slotRepository.findByIdAndEventId(slotId, event.getId())
-                                .orElseThrow(() -> new AppException("Slot not found with slotId and eventId", HttpStatus.NOT_FOUND, ErrorCode.SLOT_NOT_FOUND));
+                                .orElseThrow(() -> new AppException("Slot not found with slotId and eventId",
+                                                HttpStatus.NOT_FOUND, ErrorCode.SLOT_NOT_FOUND));
 
                 slotValidator.validate(event.getEventType(), slotRequestDto);
 
@@ -184,4 +220,26 @@ public class SlotServiceImpl implements SlotService {
 
                 return dtoMapper.toSlotResponseDto(updatedSlot);
         }
+
+        @Override
+        public DeleteValidationResponseDto slotDeleteValidation(Long eventId, Long slotId) {
+                User user = userRepository.findById(this.securityUtil.requireUserId())
+                                .orElseThrow(() -> new AppException("User not found", HttpStatus.NOT_FOUND,
+                                                ErrorCode.USER_NOT_FOUND));
+                eventRepository.findByIdAndUser(eventId, user)
+                                .orElseThrow(() -> new AppException("Event not found with eventId and User",
+                                                HttpStatus.NOT_FOUND, ErrorCode.EVENT_NOT_FOUND));
+
+                slotRepository.findByIdAndEventId(slotId, eventId)
+                                .orElseThrow(() -> new AppException("Slot not found with slotId and eventId",
+                                                HttpStatus.NOT_FOUND, ErrorCode.SLOT_NOT_FOUND));
+
+                List<Booking> bookingList = bookingRepository.findBySlot_IdAndSlot_Event_IdAndIsDeletedFalse(slotId, eventId);
+
+                DeleteValidationResponseDto responseDto = deleteValidationService
+                                .buildDeleteValidationResponseDto(bookingList);
+
+                return responseDto;
+        }
+
 }
